feat(US-006): wire real search command — all four modes, --where, --since, scope priority
Replace emit_search_stub with a fully wired handler calling run_search() from
talon-core/src/query/search.rs. The handler supports hybrid (default), semantic,
fulltext, and title modes; applies --where frontmatter filters and --since
timestamp post-filtering; multiplies scores by ScopePriority::multiplier(); and
skips expansion + rerank when --fast is set.

Also:
- Remove SearchResponse::empty_scaffold (no remaining call sites)
- Change SearchResult.score / raw_score from f32 to f64 to eliminate cast_possible_truncation
- Fix 22 clippy violations across search.rs, command.rs, cli.rs, output.rs
- Add 6-test integration suite in tests/search_integration.rs covering hybrid,
  fulltext, title, --where, --since, and empty-query paths
- 256 tests pass

+++ crates/talon-cli/src/output.rs
@@ -238,7 +238,7 @@ struct AgentSearchHit<'a> {
     path: &'a str,
     title: &'a str,
     snippet: &'a str,
-    score: f32,
+    score: f64,
 }
 
 impl<'a> From<&'a SearchResult> for AgentSearchHit<'a> {
